requirements.txt
torch==2.2.0
torchvision==0.17.0
torchaudio==2.2.0
torchmetrics
numpy==1.26.4
tqdm
matplotlib
pandas
wandb
comet_ml
hydra-core
kagglehub
pysoundfile
ffmpeg

black
isort
pre-commit
